requirements-dev.txt
# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.21.2
pytest-cov==4.1.0
httpx==0.26.0
faker==22.6.0
factory-boy==3.3.0

# Include production dependencies
-r requirements.txt